@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
-import { AbstractControl, FormControl, FormGroup, ValidationErrors, Validators } from '@angular/forms';
+import { AbstractControl, UntypedFormControl, UntypedFormGroup, ValidationErrors, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AuthenticationService, User } from 'src/app/services/authentication.service';
 
@@ -13,11 +13,11 @@ export class OpnieuwInstellenPageComponent implements OnInit {
   isBusy: boolean = false;
   errorMessage: string | null = null;
   
-  formGroup = new FormGroup({
-    wachtwoord1: new FormControl('', [
+  formGroup = new UntypedFormGroup({
+    wachtwoord1: new UntypedFormControl('', [
       Validators.required
     ]),
-    wachtwoord2: new FormControl('', [
+    wachtwoord2: new UntypedFormControl('', [
       Validators.required,
     ])
   }, [
@@ -59,7 +59,7 @@ export class OpnieuwInstellenPageComponent implements OnInit {
 }
 
 function passwordsAreEqual(control: AbstractControl): ValidationErrors | null {
-  const group = control as FormGroup;
+  const group = control as UntypedFormGroup;
   if(group.controls.wachtwoord1.value !== group.controls.wachtwoord2.value) {
     return {
       "passwordsAreEqual": true
